+++ common_test.go
@@ -2,7 +2,6 @@ package dvara
 
 import (
 	"os"
-	"sync"
 	"testing"
 	"time"
 
@@ -37,6 +36,7 @@ type Harness struct {
 	T          testing.TB
 	Stopper    stopper // This is either mgotest.Server or mgotest.ReplicaSet
 	ReplicaSet *ReplicaSet
+	Manager    *StateManager
 	Graph      *inject.Graph
 	Log        nopLogger
 }
@@ -55,13 +55,14 @@ func newHarnessInternal(url string, s stopper, t testing.TB) *Harness {
 		MaxPerClientConnections: 250,
 		GetLastErrorTimeout:     5 * time.Minute,
 		MessageTimeout:          5 * time.Second,
-		Mutex:                   &sync.RWMutex{},
 	}
+	manager := NewStateManager(&replicaSet)
 	var log nopLogger
 	var graph inject.Graph
 	err := graph.Provide(
 		&inject.Object{Value: &log},
 		&inject.Object{Value: &replicaSet},
+		&inject.Object{Value: manager},
 		&inject.Object{Value: &stats.HookClient{}},
 	)
 	ensure.Nil(t, err)
@@ -78,6 +79,7 @@ func newHarnessInternal(url string, s stopper, t testing.TB) *Harness {
 		T:          t,
 		Stopper:    s,
 		ReplicaSet: &replicaSet,
+		Manager:    manager,
 		Graph:      &graph,
 	}
 }
@@ -117,11 +119,11 @@ func (h *Harness) Stop() {
 }
 
 func (h *Harness) ProxySession() *mgo.Session {
-	return h.Dial(h.ReplicaSet.ProxyMembers()[0])
+	return h.Dial(h.Manager.ProxyMembers()[0])
 }
 
 func (h *Harness) RealSession() *mgo.Session {
-	return h.Dial(h.ReplicaSet.lastState.Addrs()[0])
+	return h.Dial(h.Manager.currentReplicaSetState.Addrs()[0])
 }
 
 func (h *Harness) Dial(u string) *mgo.Session {
