+++ healthcheck_test.go
@@ -7,8 +7,8 @@ import (
 )
 
 type FakeReplicaSet struct {
-	restartWasCalled  bool
-	CheckReturnsError bool
+	handleFailureCalled bool
+	CheckReturnsError   bool
 }
 
 func (frs *FakeReplicaSet) Check() error {
@@ -18,8 +18,8 @@ func (frs *FakeReplicaSet) Check() error {
 	return nil
 }
 
-func (frs *FakeReplicaSet) RestartIfFailed() {
-	frs.restartWasCalled = true
+func (frs *FakeReplicaSet) HandleFailure() {
+	frs.handleFailureCalled = true
 }
 
 func TestEnsureRestartIsCalled(t *testing.T) {
@@ -35,7 +35,7 @@ func TestEnsureRestartIsCalled(t *testing.T) {
 	time.Sleep(5 * time.Millisecond)
 	hc.Cancel = true
 
-	if frs.restartWasCalled == false {
+	if frs.handleFailureCalled == false {
 		t.Fatalf("Restart function not called :( %s", frs)
 	}
 
@@ -54,7 +54,7 @@ func TestEnsureRestartIsNotCalled(t *testing.T) {
 	time.Sleep(5 * time.Millisecond)
 	hc.Cancel = true
 
-	if frs.restartWasCalled == true {
+	if frs.handleFailureCalled == true {
 		t.Fatalf("Restart function not called :( %s", frs)
 	}
